requirements.txt
absl-py==0.9.0
apipkg==1.5
appdirs==1.4.4
attrs==19.3.0
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0.post1
coverage==5.2.1
cycler==0.10.0
Cython==0.29.21
Deprecated==1.2.10
distlib==0.3.1
distro==1.5.0
fonttools==4.13.0
fs==2.4.11
html5lib==1.1
idna==2.10
importlib-metadata==1.6.1
iniconfig==1.0.1
kiwisolver==1.2.0
matplotlib==3.3.1
more-itertools==8.4.0
msgpack==1.0.0
numpy==1.19.1
ordered-set==4.0.2
packaging==20.4
pep517==0.8.2
Pillow==7.2.0
pluggy==0.13.1
progress==1.5
protobuf==3.12.4
pulsemixer==1.5.1
py==1.9.0
pydub==0.24.1
PyGithub==1.52
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
resolvelib==0.4.0
retrying==1.3.3
six==1.15.0
team==1.0
toml==0.10.1
unicodedata2==13.0.0.post2
urllib3==1.25.10
webencodings==0.5.1
wrapt==1.12.1
youtube-dl==2020.7.28
zipp==3.1.0
